Merits and Flaws block encapsulation Fixes #485
